+++ CHANGES.md
@@ -40,6 +40,8 @@ unreleased
 
 - Add `--trace-file` option to trace dune internals (#1639, fix #1180, @emillon)
 
+- Remove `--stats` and track fd usage in `--trace-file` (#1667, @emillon)
+
 1.6.2 (05/12/2018)
 ------------------